DAT-802 refactor error module
