--- app/src/app/error/components/error-forbidden/error-forbidden.component.html
@@ -1,10 +1,0 @@
-<div class="page-wrap d-flex flex-row align-items-center">
-  <div class="container">
-    <div class="row justify-content-center">
-      <div class="col-md-12 text-center">
-        <span class="display-1 d-block">403</span>
-        <div class="mb-4 lead">The page or resource you were trying to read is absolutely forbidden for some reason.</div>
-      </div>
-    </div>
-  </div>
-</div>
