--- app/src/app/error/components/error-forbidden/error-forbidden.component.ts
@@ -1,14 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  templateUrl: './error-forbidden.component.html',
-  styleUrls: ['./error-forbidden.component.scss']
-})
-export class ErrorForbiddenComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
