+++ app/src/app/modules/error/components/error-forbidden/error-forbidden.component.ts
@@ -11,15 +11,15 @@
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
-
+*
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
+import { Component } from '@angular/core';
+/* ------------------------------------------------------------------------------------------------------------------ */
 
-export class BackendHttpError {
-  public status: number;
-  public response: string;
-  public description: string;
-  public message: string;
-  public joke?: string;
-}
+@Component({
+    templateUrl: './error-forbidden.component.html',
+    styleUrls: ['./error-forbidden.component.scss']
+})
+export class ErrorForbiddenComponent {}
