+++ app/src/app/modules/error/components/error-not-found/error-not-found.component.ts
@@ -11,36 +11,15 @@
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
-
+*
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
+import { Component } from '@angular/core';
 
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { ErrorComponent } from './error.component';
-
-const routes: Routes = [
-  {
-    path: '',
-    pathMatch: 'full',
-    data: {
-      breadcrumb: 'Not found'
-    },
-    redirectTo: 'dashboard'
-  },
-  {
-    path: ':statusCode',
-    data: {
-      breadcrumb: 'Status'
-    },
-    component: ErrorComponent
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
+@Component({
+    selector: 'cmdb-error-not-found',
+    templateUrl: './error-not-found.component.html',
+    styleUrls: ['./error-not-found.component.scss']
 })
-export class ErrorRoutingModule {
-}
+export class ErrorNotFoundComponent {}
