+++ app/src/app/modules/error/error.list.ts
@@ -11,15 +11,15 @@
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
-
+*
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-
 import { ErrorForbiddenComponent } from './components/error-forbidden/error-forbidden.component';
 import { ErrorNotFoundComponent } from './components/error-not-found/error-not-found.component';
+/* ------------------------------------------------------------------------------------------------------------------ */
 
 export const errorComponents: { [code: number]: any } = {
-  403: ErrorForbiddenComponent,
-  404: ErrorNotFoundComponent
+    403: ErrorForbiddenComponent,
+    404: ErrorNotFoundComponent
 };
