+++ app/src/app/modules/error/error.module.ts
@@ -19,7 +19,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { ErrorRoutingModule } from './error-routing.module';
-import { LayoutModule } from '../layout/layout.module';
+import { LayoutModule } from '../../layout/layout.module';
 
 import { ErrorComponent } from './error.component';
 import { ErrorForbiddenComponent } from './components/error-forbidden/error-forbidden.component';
@@ -41,5 +41,4 @@ import { ErrorNotFoundComponent } from './components/error-not-found/error-not-f
         ErrorNotFoundComponent,
     ]
 })
-export class ErrorModule {
-}
+export class ErrorModule {}
